+++ packages/expo/src/devtools/getConnectionInfo.native.ts
@@ -5,7 +5,8 @@
 import type { ConnectionInfo } from './devtools.types';
 
 export function getConnectionInfo(): Omit<ConnectionInfo, 'pluginName'> {
-  const getDevServer = require('react-native/Libraries/Core/Devtools/getDevServer').default;
+  const devServerModule = require('react-native/Libraries/Core/Devtools/getDevServer');
+  const getDevServer = devServerModule.default ?? devServerModule;
   const devServer = getDevServer()
     .url.replace(/^https?:\/\//, '')
     .replace(/\/?$/, '') as string;